Add a circuit node interface (#99)
This pull request adds:
- A new class: CircuitNodeId 
  - A CircuitNodeIds is a combination of a population and a node id. --> ("population_name", node_id). 
- This implies to also allow the sequence of circuit node id as queries. So you have a low performance container using list of ids for the queries. We also provide a "high performance" one : CircuitNodeIds.
- A new class : CircuitNodeIds, based on pandas.MultiIndex to store and manipulate of global circuit ids. This is a high performance container for queries. I used composition and not heritage for this class because all the useful classmethods from pandas.MultiIndex return MutiIndex and not cls(). Therefore, implementing the `\_\_call\_\_` allows to use it as an index in the dataframes.

I chose this over a simple offsetting of ids because this is more compatible with the sonata main feature of "adding extra populations" to the circuit. With this design, the direct calls to ids in analysis scripts or the "on-disk stored" node ids (more common than we think ...) are not invalidated with new populations. We can also trustfully compare two sets of ids which is not the case with offsets. And since we allow people to use the sonata node ids or global ids this reduce the possible confusion and more flexibility for the API. I also don t want people to mix the sonata node ids and the global ids.
This also allows to do simple operations more transparently on ids like sorting/append/comparisons/filtering without relying on the circuit itself or mapping. 


Ex of uses for this class:
```
>>> ids_pop12 = CircuitNodeIds.from_arrays(["pop1", "pop1", "pop2"], [0, 1, 0])  # main 
[("pop1", 0), ("pop1", 1), ("pop2", 0)]
>>> ids_pop3 = CircuitNodeIds.from_dict({"pop3": [0, 1]})
[("pop3", 0), ("pop3", 1)]
ids_pop123 = ids_pop12.append(ids_pop3)
ids_pop1 = ids_pop123.filter_population("pop1")
sample_ids_pop123 = ids_pop123.sample(2)
```

- A new class: Nodes. This class aims at : 
  - Accessing the different node populations (like the current `circuit.nodes` does).
  - For compat you can iterate on the population using the dict interface : items(), values(), keys() or the wrapper populations(), names().
  - Allowing queries on the whole circuit for nodes. You can now do `circuit.nodes.ids(query)` and all the nodes from all populations matching the query will be returned. This should be handy for the spacial queries on different type of cells (like glia cells mixed with neurons) or for the combination of multiple regions merged together containing the same type of cells (like neocortex split in multiple sub region). 
  - Allowing queries on the whole circuit for nodes properties. You can now do `circuit.nodes.get(query, properties=[prop1, prop2])`. This will return a dataframe indexed by a MultiIndex (population, id). You will be able to query on this dataframe using : 
tuple ("population", id), CircuitNodeId, a list of tuples or a list of CircuitNodeId or a MultiIndex or CircuitNodeIds : 
```
ids = circuit.nodes.ids(query)  #  CircuitNodeIds corresponding to the query
data = circuit.nodes.get() # all data from all populations 
data.loc[ids]  # identical to circuit.nodes.get(query)
```
if a property is missing for a node id (this is possible due to population mixing) a NaN is attributed to the missing property value.

- The edges and morph class has been updated. You can use the CircuitNodeId/CircuitNodeIds inside the edges functions and the CircuitNodeId inside MorphHelper.
- Tests has been added accordingly.