Add include_empty keyword argument
When using get-like functions (edges.pathway_edges, nodes.get) it is common to concatenate
the outputs of the functions to mimic the old bluepysnap interface.
However, if no nodes/edges satisfy the query, instead of resulting in an empty dataframe,
this approach will result in an error, because no dataframes at all are yielded by the generator.

The include_empty keyword argument allows the user to specify that they still want to see a dataframe if the query turns up nothing.

One example where this is useful for me is when I use the length of a connection dataframe to determine connection probability.